First page Info updating (Todo) small problem with DB consistency-> Not confirmed to be solved
