+++ app/src/main/java/org/tensorflow/lite/examples/objectdetection/QrCodeViewModel.kt
@@ -48,6 +48,8 @@ class QrCodeViewModel(barcode: Barcode) {
                             MyEntryPoint.prefs.setString("Date", words[3])
                             MyEntryPoint.prefs.setString("hash", words[4])
 
+                            val cameraIntent = Intent(v.context, CameraActivity::class.java)
+                            v.context.startActivity(cameraIntent)
                         }
                         else{
                             qrContent = "Cannot parse: $qrContent"
